+++ modules/apps/foundation/frontend-editor/frontend-editor-alloyeditor-web/build.gradle
@@ -65,5 +65,5 @@ dependencies {
 	provided group: "javax.servlet.jsp", name: "javax.servlet.jsp-api", version: "2.3.1"
 	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	provided project(":apps:foundation:petra:petra-encryptor-util")
+	provided project(":apps:foundation:petra:petra-encryptor")
 }
