+++ modules/apps/foundation/petra/petra-encryptor/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay Petra Encryptor Utilities
-Bundle-SymbolicName: com.liferay.petra.encryptor.util
+Bundle-SymbolicName: com.liferay.petra.encryptor
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.petra.encryptor.util
+Export-Package: com.liferay.petra.encryptor
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title:
