+++ modules/apps/foundation/petra/petra-encryptor/src/main/java/com/liferay/petra/encryptor/EncryptorException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.encryptor.util;
+package com.liferay.petra.encryptor;
 
 /**
  * @author Brian Wing Shun Chan
