+++ modules/apps/foundation/portal-osgi-web/portal-osgi-web-portlet-container-test/build.gradle
@@ -8,7 +8,7 @@ dependencies {
 	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.util.java", version: "2.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.util.taglib", version: "2.0.0"
-	provided project(":apps:foundation:petra:petra-encryptor-util")
+	provided project(":apps:foundation:petra:petra-encryptor")
 
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.arquillian.extension.junit.bridge", version: "1.0.5"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
