+++ util-java/src/com/liferay/util/EncryptorException.java
@@ -16,7 +16,7 @@ package com.liferay.util;
 
 /**
  * @author Brian Wing Shun Chan
- * @see com.liferay.petra.encryptor.util.EncryptorException
+ * @see com.liferay.petra.encryptor.EncryptorException
  */
 public class EncryptorException extends Exception {
